Move pending_operations.py to separate folder
Hexagonal
