+++ datastore_version_manager/domain/pending_operations.py
